Add NTP clock correction
Add code to generalize the UDP library code.  Extract the NTP code
from its origin so it can be used generically with this UDP abstraction.

The result works fine on the PC but fails to compile for Arduino.
But the failure seems unrelated to the code; it appears to be a
missing include file definition or something like that.