Rename .ino -> .cpp except the one main file
The Arduino IDE seems to accept the .cpp files as easily
as the .ino files.  This change makes the main .ino file
easier for users to identify.  Maybe it also prevents the
IDE from concatenating files together into one, but
I didn't check.  Anyway, we do not need that anymore.