Move files to appease arduino IDE
The Arduino IDE insists that a .ino file be stored in a folder
bearing the same name.  This project is named 'master_clock' and
the .ino file is named 'master_clock.ino', so this has worked for
me so far.  But when the .zip file is downloaded from github the
top-level folder is renamed slightly.  Also a user may rename the
project folder when the project is cloned thus making the project
unbuildable via the IDE.

Move the arduino files into a subfolder named 'master_clock' to
match the master_clock.ino filename.  Adjust the makefiles and
.gitignore files to match the new layout.

Also fix the .ino file to include the library headers so the IDE
will notice the paths are needed when compiling the .cpp files.