Add a telnet console server on port 23
When we monitor the clock over the serial port, the arduino resets
to allow for a new code upload.  This is a useful feature, but it
means we cannot easily connect to the UART to check status when
things go wrong since the act of connecting will reset our status
anyway.

Add a single-client telnet server so we can telnet to the clock
to monitor it over ethernet instead of over the UART.

Currently this interface is not implemented in the PC simulator
version.  There are two complications to adding it there.  One
is that port 23 is privileged and would require root access. This
can be overcome by using a different port (2323) which is not
privileged.

The other complication is simply that I have not written the
Ethernet code needed for the PC side yet.
other is that