Add generic ConfigData handlers
Create a base class to hold a "configurable item" and then derive
specific type instances of it.  Create objects to hold various
configurable and reportable data for the master clock.  Report the
config data to the console when connected.

Prepare to report the same data via XML on the WebServer, to
store/load the config data to/from EEPROM (or local file), and
to permit the user to read/write the config via SET and GET
commands, for example.