Teach Time class to notice when NTP is synchronized
When we can't reach NTP or the clock is seriously wrong, don't
bother following the system reported localtime because it's probably
wrong. Use a fake local clock based on monotonic() and the time we
believe is displayed on the clock.  Check again for NTP sync every
5 minutes until we get one.  Then assume we are synced until 24 hours
goes by.  This should keep our load down on the NTP servers.

Also, add a subtract operator to the time class that returns a number
of seconds between two times. The real number is always positive and
represents the number of seconds until one time catches up with the
other on a 12-hour clock.