Add catch-up / wait modes for FAST or SLOW times
If we know the clock is fast by a few minutes, wait for time to
catch up instead of trying to go all the way around the clock face for
12 hours.

If we know we're slow, force the pulses and bump the face time a whole
minute so we get to the next minute without waiting on the Simplex
protocol to catch mistakes.  This assumes we're reliable enough that we
won't have many pulse errors in the catch-up race.  We can tolerate some
because we'll correct on the next hour, but we can't tolerate more than 10%
errors, I guess.

Slow down in power-lost mode to save battery.