Overhaul time tracking to use internal RTC
Separate the "real time" from the "displayed time" so we can detect
differences of more than one hour.  Collapse the displayed time into
a single "minutes" value called walltime. Whenever the walltime is
different from the realtime, adjust the walltime until it matches
Remove all the separate m/s trackers and rely on the delta between
realtime and walltime to tell us when to move the hands.

Remove some of the console code that used to take advantage of the
manual timekeeping.

Move run_switch detection out into the markTime routine that syncs
the internal and external clocks.