Simplify console showTimer to be self-regulating
showTime() used to be called regularly once per second because
of how we did other tasks.  Now we do things differently so we
need to regulate ourselves to only call it once per second. This
causes weird races when we get out of sync with other processes.
Let showTime regulate itself so we can just call it all the damn time.