add indicator to report untrusted clockface-time
We record the clockface time periodically so we can remember it
on reboot. But if the dat file is corrupted or missing, we just assume
the clockface is the current RTC time so we can start advancing the
clock minute-by-minute in "dumb" mode.

On startup, though, the RTC clock may be wrong. The pi tries to restore
some "reasonable" time it, it seems, but the ntp sync will tell us the real
story later on.  In the old code we would assume the RTC was the same
as the clockface time and then start using it.  Later, when we synced and
got the real time, we could be way off from the original assumption,
causing us to go nuts advancing the clockface from what we thought it
was (which was wrong) to what it should be (still wrong, since we don't
know where we started).

Add a "trusted" flag to the SimplexProtocol that indicates whether
our clockface time is trusted or not.  Set trusted to False until we
successfully load a time value from a file or we sync from the NTP
server. That is, if we can't know the clockface time from the file,
assume it matches the real NIST time, even though we don't know
that value yet.  When we finally do know the real NIST time, adjust
the clockface conceptual time to match and assume it had matched
all along.  This is probably wrong, of course, but if we are within 30
minutes of correct, we will auto-adjust soon via the Simplex protocol.